+++ packages/jet/lib/jet_framework.dart
@@ -5,25 +5,47 @@ export 'bootstrap/boot.dart';
 export 'resources/theme/notifiers/theme_switcher_notifier.dart';
 export 'localization/widgets/language_switcher.dart';
 export 'resources/theme/widgets/theme_switcher.dart';
+export 'resources/state/jet_builder.dart';
 export 'localization/i10n.dart';
 export 'storage/local_storage.dart';
 export 'networking/networking.dart';
 export 'package:auto_route/auto_route.dart';
 export 'package:hooks_riverpod/hooks_riverpod.dart';
-export 'package:flutter_form_builder/flutter_form_builder.dart';
+export 'package:flutter_hooks/flutter_hooks.dart';
 export 'helpers/helpers.dart';
 export 'package:flutter_screenutil/flutter_screenutil.dart';
-export 'package:form_builder_validators/form_builder_validators.dart';
 export 'package:phosphor_flutter/phosphor_flutter.dart';
 export 'package:guardo/guardo.dart';
 
+// Flutter Form Builder - only export necessary types
+export 'package:flutter_form_builder/flutter_form_builder.dart'
+    show
+        FormBuilderState,
+        FormFieldValidator,
+        FormBuilderTextField,
+        FormBuilderDropdown,
+        FormBuilderCheckbox,
+        FormBuilderSwitch,
+        FormBuilderDateTimePicker,
+        InputType,
+        FormBuilderField,
+        DropdownMenuItem,
+        ValueTransformer,
+        AutovalidateMode;
+
 // Widgets
 export 'widgets/widgets.dart';
+
+// Components
+export 'resources/components/components.dart';
+export 'package:smooth_page_indicator/smooth_page_indicator.dart';
 
 // Forms
 export '../forms/common.dart';
 export '../forms/notifiers/jet_form_notifier.dart';
 export '../forms/widgets/jet_form_builder.dart';
+export '../forms/widgets/jet_form_widget.dart';
+export '../forms/validators/jet_validators.dart';
 export '../forms/inputs/inputs.dart';
 
 // Error Handling